json_path test cleanup
